BCM2708 PWM driver update.
Filled in some of the code body for the BCM2708 PWM driver.
This sort of does the right thing, but limited debugging on the
physical pin inconclusive.  Added PWM and CPRMAN I/O memory
regions to BCM2708 I/O map.  Added routines in GPIO code to
allow other drivers (PWM) to switch pins to alternative functions.